+++ apps/doc/src/lib/types/navs.ts
@@ -1,12 +1,14 @@
+import type {UiAssetType} from '@fat-fuzzy/ui'
+
 export type NavItem = {
 	slug: string
 	title: string
 	label: string
 	asset?: string
-	assetType?: string
-	reveal?: string
+	assetType?: UiAssetType
 	formaction?: string
 	actionPath?: string
 	layout?: string
 	items?: NavItem[]
+	url?: URL
 }