+++ apps/doc/src/routes/(rails)/+layout.server.ts
@@ -1,6 +1,4 @@
-import ui from '@fat-fuzzy/ui'
 import {buildNav} from '$data/nav'
-const {DEFAULT_PREFERENCES, DEFAULT_REVEAL_STATE} = ui.constants
 
 export const load = async ({locals, url}) => {
 	const pageName = url.pathname.split('/')
@@ -8,7 +6,6 @@ export const load = async ({locals, url}) => {
 
 	if (pageName.length > 0 && pageName[1]) {
 		sidebar = buildNav(pageName[1])
-		sidebar.reveal = locals.sidebar.reveal ?? sidebar.reveal
 		sidebar.actionPath = url.pathname
 	}
 
@@ -16,17 +13,12 @@ export const load = async ({locals, url}) => {
 		sidebar = {layout: 'tgv', items: []}
 	}
 
-	const appContext = locals.appContext ?? {
-		...DEFAULT_REVEAL_STATE,
-		...DEFAULT_PREFERENCES,
-	}
-	const pageContext = locals.pageContext
-	appContext.actionPath = url.pathname
+	const pageContext = locals.pageContext ?? {}
+
 	pageContext.actionPath = url.pathname
 
 	return {
 		sidebar,
 		pageContext,
-		appContext,
 	}
 }