+++ packages/ui/src/lib/components/recipes/tabs/Tabs.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type {TabsProps, Tab} from '$types'
-	import styleHelper from '$lib/utils/styles.js'
+	import styleHelper from '$lib/utils/styles'
 	import {onMount} from 'svelte'
 
 	let {