+++ packages/ui/src/lib/types/constants.ts
@@ -1,8 +1,8 @@
 import type {
 	AppContextItems,
-	UiRevealState,
 	CookiePreferences,
 	ViewingPreferences,
+	PrivacyPreferences,
 	NavItem,
 } from '$types'
 
@@ -17,16 +17,12 @@ const DEFAULT_COOKIES_CONSENT: CookiePreferences = {
 	analytics: false,
 }
 
-const DEFAULT_REVEAL_STATE: UiRevealState = {reveal: 'collapsed'}
-
-const DEFAULT_SIDEBAR_REVEAL_STATE: UiRevealState = {reveal: 'collapsed'}
-
-const DEFAULT_NAV_REVEAL_STATE: UiRevealState = {reveal: 'collapsed'}
-
 const DEFAULT_PREFERENCES: ViewingPreferences = {
-	reveal: 'collapsed',
 	brightness: 'day',
 	contrast: 'contrast',
+}
+
+const DEFAULT_COOKIES_PREFERENCES: PrivacyPreferences = {
 	consent: DEFAULT_COOKIES_CONSENT,
 }
 
@@ -183,10 +179,8 @@ export default {
 	UI_STATE,
 	STATE_SWITCHER,
 	DEFAULT_STYLES,
-	DEFAULT_REVEAL_STATE,
-	DEFAULT_NAV_REVEAL_STATE,
-	DEFAULT_SIDEBAR_REVEAL_STATE,
 	DEFAULT_PREFERENCES,
+	DEFAULT_COOKIES_PREFERENCES,
 	DEFAULT_SCENE_ID,
 	TRANSITION_REVEAL,
 	TRANSITION_BRIGHTNESS,