+++ packages/ui/tests/browser/utils/ClickOutsideTest.svelte
@@ -9,6 +9,9 @@
 </script>
 
 <div>
+	<!-- 
+// TODO: update : use {@attach clickOutside}
+  -->
 	<p
 		use:clickOutside={{callback: handleClickOutside, ignore: ['ignored']}}
 		class="click-outside"